pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-rag-sales"
version = "0.1.0"
description = "Agentic RAG system for AI real estate sales assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]",
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "qdrant-client>=1.9.0",
    "httpx>=0.27.0",
    "redis>=5.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "numpy<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
